# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GreenSync Lawn Care SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List routes
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteListResponse"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Get route with stops
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
    put:
      operationId: updateRoute
      tags: [routes]
      summary: Update route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List automation rules
      responses:
        "200":
          description: Automation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationListResponse"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /automations/{id}/toggle:
    post:
      operationId: toggleAutomation
      tags: [automations]
      summary: Toggle automation active state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"

  # ─── TEAM ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH SCHEMAS ──────────────────────────────────────────────────────────
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        customerId:
          type: integer
        propertyId:
          type: integer
          nullable: true
        serviceId:
          type: integer
          nullable: true
        assignedUserId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
          nullable: true
        price:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        internalNotes:
          type: string
          nullable: true
        completionNotes:
          type: string
          nullable: true
        reminderSent:
          type: boolean
        customerName:
          type: string
          nullable: true
        serviceName:
          type: string
          nullable: true
        assignedUserName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, customerId, status, scheduledStart, reminderSent, createdAt]
...
    Route:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
          nullable: true
        routeDate:
          type: string
          format: date-time
        assignedUserId:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        stopsCount:
          type: integer
          nullable: true
        assignedUserName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, routeDate, createdAt]
    RouteStop:
      type: object
      properties:
        id:
          type: integer
        routeId:
          type: integer
        appointmentId:
          type: integer
        stopOrder:
          type: integer
        estimatedArrival:
          type: string
          format: date-time
          nullable: true
        appointment:
          $ref: "#/components/schemas/Appointment"
      required: [id, routeId, appointmentId, stopOrder]
    RouteDetail:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/Route"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
      required: [route, stops]
    RouteListResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/Route"
      required: [routes]
    CreateRouteRequest:
      type: object
      required: [routeDate]
      properties:
        name:
          type: string
        routeDate:
          type: string
          format: date-time
        assignedUserId:
          type: integer
        status:
          type: string
        notes:
          type: string
...
    Automation:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        triggerType:
          type: string
        actionType:
          type: string
        configJson:
          type: object
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, triggerType, actionType, isActive, createdAt]
    AutomationListResponse:
      type: object
      properties:
        automations:
          type: array
          items:
            $ref: "#/components/schemas/Automation"
      required: [automations]
    CreateAutomationRequest:
      type: object
      required: [name, triggerType, actionType]
      properties:
        name:
          type: string
        triggerType:
          type: string
        actionType:
          type: string
        configJson:
          type: object
        isActive:
          type: boolean

    # ─── TEAM SCHEMAS ──────────────────────────────────────────────────────────
...